Fix all broken test cases
A number of test cases failed, possibly due to changes in Ensemble. All
test cases have been made consistent with the data that is *currently*
being returned from Ensemble. To prevent these kind of issues in the
future, all SNP id's which caused problems have been added to the lookup
file, so that no Ensemble lookups are required to run the test cases.